+++ viajes/lib/home/home_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
-import 'package:viajes/home/icon_button.dart';
+import 'package:viajes/home/custom_icon_button.dart';
 import 'package:viajes/home/location_card.dart';
 import 'package:viajes/nav_bar.dart';
 
@@ -144,10 +144,12 @@ class _HomePageState extends State<HomePage> {
       itemCount: 9,
       itemBuilder: (BuildContext context, int index) {
         return CustomIconButton(
-          iconID: 69690,
+          iconID: 61668,
+          iconFamily: 'FontAwesomeSolid',
+          buttonSize: 60,
           iconColor: Theme.of(context).iconTheme.color,
           buttonColor: Colors.amberAccent[100],
-          buttonPressed: () {
+          onButtonPressed: () {
             print('alo');
           },
         );
